fix imports for separate command files
