+++ src/pipx/commands/__init__.py
@@ -1,16 +1,13 @@
+from .ensure_path import ensure_path
+from .inject import inject
+from .install import install
 from .list_packages import list_packages
+from .reinstall import reinstall_all
 from .run import run
+from .run_pip import run_pip
+from .uninstall import uninstall, uninstall_all
 from .upgrade import upgrade, upgrade_all
 
-from .commands import (
-    install,
-    inject,
-    uninstall,
-    uninstall_all,
-    reinstall_all,
-    run_pip,
-    ensurepath,
-)
 
 __all__ = [
     "upgrade",
@@ -23,5 +20,5 @@ __all__ = [
     "reinstall_all",
     "list_packages",
     "run_pip",
-    "ensurepath",
+    "ensure_path",
 ]
